package diff is moved to gov.ca.cwds.idm.service
