+++ src/main/java/gov/ca/cwds/idm/service/cognito/attribute/DatabaseDiffMapBuilder.java
@@ -4,8 +4,8 @@ import static gov.ca.cwds.idm.service.cognito.attribute.DatabaseUserAttribute.NO
 
 import gov.ca.cwds.idm.dto.User;
 import gov.ca.cwds.idm.dto.UserUpdate;
-import gov.ca.cwds.idm.service.cognito.attribute.diff.Diff;
-import gov.ca.cwds.idm.service.cognito.attribute.diff.StringDiff;
+import gov.ca.cwds.idm.service.diff.Diff;
+import gov.ca.cwds.idm.service.diff.StringDiff;
 import java.util.HashMap;
 import java.util.Map;
