+++ src/main/java/gov/ca/cwds/idm/service/diff/Diff.java
@@ -1,4 +1,4 @@
-package gov.ca.cwds.idm.service.cognito.attribute.diff;
+package gov.ca.cwds.idm.service.diff;
 
 public interface Diff<T> {
