+++ src/main/java/gov/ca/cwds/idm/service/diff/RolesUserAttributeDiff.java
@@ -1,4 +1,4 @@
-package gov.ca.cwds.idm.service.cognito.attribute.diff;
+package gov.ca.cwds.idm.service.diff;
 
 import static gov.ca.cwds.config.api.idm.Roles.replaceRoleIdByName;
