+++ src/main/java/gov/ca/cwds/idm/service/diff/builder/CollectionAttributeDiffBuilder.java
@@ -1,8 +1,8 @@
-package gov.ca.cwds.idm.service.cognito.attribute.diff.builder;
+package gov.ca.cwds.idm.service.diff.builder;
 
 import gov.ca.cwds.idm.service.cognito.attribute.UserAttribute;
-import gov.ca.cwds.idm.service.cognito.attribute.diff.CollectionUserAttributeDiff;
-import gov.ca.cwds.idm.service.cognito.attribute.diff.UserAttributeDiff;
+import gov.ca.cwds.idm.service.diff.CollectionUserAttributeDiff;
+import gov.ca.cwds.idm.service.diff.UserAttributeDiff;
 import java.util.Set;
 
 /**
