+++ src/main/java/gov/ca/cwds/idm/service/diff/builder/RoleAttributeDiffBuilder.java
@@ -1,9 +1,8 @@
-package gov.ca.cwds.idm.service.cognito.attribute.diff.builder;
+package gov.ca.cwds.idm.service.diff.builder;
 
-import gov.ca.cwds.idm.dto.User;
 import gov.ca.cwds.idm.service.cognito.attribute.CustomUserAttribute;
-import gov.ca.cwds.idm.service.cognito.attribute.diff.RolesUserAttributeDiff;
-import gov.ca.cwds.idm.service.cognito.attribute.diff.UserAttributeDiff;
+import gov.ca.cwds.idm.service.diff.RolesUserAttributeDiff;
+import gov.ca.cwds.idm.service.diff.UserAttributeDiff;
 import java.util.Set;
 
 /**
