+++ src/main/java/gov/ca/cwds/idm/service/diff/builder/StringAttributeDiffBuilder.java
@@ -1,8 +1,8 @@
-package gov.ca.cwds.idm.service.cognito.attribute.diff.builder;
+package gov.ca.cwds.idm.service.diff.builder;
 
 import gov.ca.cwds.idm.service.cognito.attribute.UserAttribute;
-import gov.ca.cwds.idm.service.cognito.attribute.diff.StringUserAttributeDiff;
-import gov.ca.cwds.idm.service.cognito.attribute.diff.UserAttributeDiff;
+import gov.ca.cwds.idm.service.diff.StringUserAttributeDiff;
+import gov.ca.cwds.idm.service.diff.UserAttributeDiff;
 
 /**
  * Created by Alexander Serbin on 1/15/2019
