+++ src/main/java/gov/ca/cwds/idm/service/diff/builder/UserAttributeDiffBuilder.java
@@ -1,6 +1,6 @@
-package gov.ca.cwds.idm.service.cognito.attribute.diff.builder;
+package gov.ca.cwds.idm.service.diff.builder;
 
-import gov.ca.cwds.idm.service.cognito.attribute.diff.UserAttributeDiff;
+import gov.ca.cwds.idm.service.diff.UserAttributeDiff;
 
 /**
  * Created by Alexander Serbin on 1/15/2019
