+++ src/test/java/gov/ca/cwds/idm/event/UserChangeLogEventTest.java
@@ -26,13 +26,13 @@ import gov.ca.cwds.idm.service.UserUpdateRequest;
 import gov.ca.cwds.idm.service.authorization.UserRolesService;
 import gov.ca.cwds.idm.service.cognito.attribute.OtherUserAttribute;
 import gov.ca.cwds.idm.service.cognito.attribute.UserAttribute;
-import gov.ca.cwds.idm.service.cognito.attribute.diff.CollectionUserAttributeDiff;
-import gov.ca.cwds.idm.service.cognito.attribute.diff.Diff;
-import gov.ca.cwds.idm.service.cognito.attribute.diff.RolesUserAttributeDiff;
-import gov.ca.cwds.idm.service.cognito.attribute.diff.StringDiff;
-import gov.ca.cwds.idm.service.cognito.attribute.diff.StringUserAttributeDiff;
-import gov.ca.cwds.idm.service.cognito.attribute.diff.UserAttributeDiff;
-import gov.ca.cwds.idm.service.cognito.attribute.diff.UserEnabledStatusAttributeDiff;
+import gov.ca.cwds.idm.service.diff.CollectionUserAttributeDiff;
+import gov.ca.cwds.idm.service.diff.Diff;
+import gov.ca.cwds.idm.service.diff.RolesUserAttributeDiff;
+import gov.ca.cwds.idm.service.diff.StringDiff;
+import gov.ca.cwds.idm.service.diff.StringUserAttributeDiff;
+import gov.ca.cwds.idm.service.diff.UserAttributeDiff;
+import gov.ca.cwds.idm.service.diff.UserEnabledStatusAttributeDiff;
 import gov.ca.cwds.util.CurrentAuthenticatedUserUtil;
 import java.util.Arrays;
 import java.util.Collections;
