+++ src/test/java/gov/ca/cwds/idm/service/NsUserBuilderTest.java
@@ -7,8 +7,8 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import gov.ca.cwds.idm.persistence.ns.entity.NsUser;
 import gov.ca.cwds.idm.service.NsUserService.NsUserBuilder;
 import gov.ca.cwds.idm.service.cognito.attribute.UserAttribute;
-import gov.ca.cwds.idm.service.cognito.attribute.diff.Diff;
-import gov.ca.cwds.idm.service.cognito.attribute.diff.StringDiff;
+import gov.ca.cwds.idm.service.diff.Diff;
+import gov.ca.cwds.idm.service.diff.StringDiff;
 import java.util.HashMap;
 import java.util.Map;
 import org.junit.Test;
